chore: move provider to utils
