+++ packages/salesforcedx-utils-vscode/src/index.ts
@@ -54,3 +54,4 @@ export * from './types';
 export * from './date';
 export * from './output';
 export * from './predicates';
+export * from './providers';
