+++ packages/salesforcedx-utils-vscode/src/providers/index.ts
@@ -1,0 +1,1 @@
+export { ConfigAggregatorProvider } from './configAggregatorProvider';
