+++ packages/salesforcedx-utils-vscode/src/workspaces/index.ts
@@ -11,3 +11,5 @@ export {
   getRootWorkspacePath,
   getRootWorkspaceSfdxPath
 } from './rootWorkspace';
+
+export { workspaceUtils } from './workspaceUtils';
