+++ packages/salesforcedx-vscode-core/src/providers/index.ts
@@ -1,0 +1,1 @@
+// export { configAggregatorProvider } from './configAggregatorProvider';
